Modificaciones en daño al jugador
Cuando el jugador recibe daño y se queda sin sus 5 vidas, vuelve a la posición inicial.
Además se modifican los tiempos de spawn de los enemigos.

+++ Proyecto Unity 2D/Assets/Scripts/Personaje/Curar.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class Curar : MonoBehaviour
 {
     [Header("Configuracion")]
-    [SerializeField] float puntos = 1f;
+    [SerializeField] int puntos = 1;
 
     private void OnParticleCollision(GameObject other)
     {
